% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_get}
\alias{error_get}
\title{Report an error when using XML with get_* fuctions}
\usage{
error_get()
}
\arguments{
\item{stations_list}{The object passed to get_* function}
}
\description{
Report an error when using XML with get_* fuctions
}
\keyword{internal}
