% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{error_status}
\alias{error_status}
\title{Report an error with a http request}
\usage{
error_status(response_text)
}
\arguments{
\item{response}{The text of the server response.}
}
\description{
Report an error with a http request
}
\keyword{internal}
