% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_authentication}
\alias{fetch_authentication}
\title{Fetch NRDP authentication token}
\usage{
fetch_authentication(nrdp_user, nrdp_pass)
}
\description{
Fetch NRDP authentication token
}
\keyword{internal}
