% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_stations_list}
\alias{fetch_stations_list}
\title{Fetch data on all stations and return as a list}
\usage{
fetch_stations_list(nrdp_user, nrdp_pass)
}
\arguments{
\item{nrdp_user}{Your National Rail Data Portal email address.}

\item{nrdp_pass}{Your National Rail Data Portal password.}
}
\value{
An R object of class "list".
}
\description{
\code{fetch_stations_list} fetches data on all stations and returns an R
object of class "list".
}
