% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_stations_xml}
\alias{fetch_stations_xml}
\title{Fetch data on all stations and return as an xml_document}
\usage{
fetch_stations_xml(nrdp_user, nrdp_pass)
}
\arguments{
\item{nrdp_user}{Your National Rail Data Portal email address.}

\item{nrdp_pass}{Your National Rail Data Portal password.}
}
\value{
An R object of class "xml_document" and "xml_node".
}
\description{
\code{fetch_stations_xml} fetches data on all stations and returns
an R object of class "xml_document" and "xml_node".
}
