% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_interchange_tags}
\alias{get_interchange_tags}
\title{Get a tibble of interchange tag data for each station}
\usage{
get_interchange_tags(stations_list)
}
\arguments{
\item{stations_list}{A list of stations as returned by \code{fetch_stations_list}}
}
\value{
A tibble of data under the interchange tag for each station.
}
\description{
\code{get_interchange_tags} takes a list of stations returned from the API
as a list and returns a tibble of data under the interchange tag, with one
row per station. Each tibble column is a character string.
}
