% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_station_tags}
\alias{get_station_tags}
\title{Get a tibble of station tag data for each station}
\usage{
get_station_tags(stations_list)
}
\arguments{
\item{stations_list}{A list of stations as returned by \code{fetch_stations_list}}
}
\value{
A tibble of data under the station tag for each station.
}
\description{
\code{get_station_tags} takes a list of stations and returns a tibble of
data under the station tag, with one row per station. Each tibble column
is a character string.
}
